backend/requirements.txt
fastapi==0.110.1
motor==3.3.1
pymongo==4.5.0
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.21
PyJWT==2.10.1
bcrypt==4.1.3
cloudinary==1.41.0
uvicorn==0.25.0
